requirements.txt
bidict==0.23.1
dnspython==2.7.0
eventlet==0.39.0
greenlet==3.1.1
h11==0.14.0
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
wsproto==1.2.0
